feat: Add new project to Projects component
This commit adds a new project to the Projects component. The new project is an e-commerce website for plastic pallets. It includes a project preview image, a project link, a title, and a description. The project preview image is updated to use the new project image, and the project link, title, and description are set accordingly.

This change addresses the need to showcase the new e-commerce website for plastic pallets in the Projects component and provide relevant information about the project to users.